tidy up dotenv support
