+++ otel/tests/phpt/dotenv/dotenv-loaded-on-rinit.phpt
@@ -19,7 +19,6 @@ Globals::tracerProvider()->getTracer('my_tracer')->spanBuilder('root')->startSpa
 ?>
 --EXPECTF--
 %A
-[%s] [DEBUG] [pid=%d] [ThreadId(%d)] OpenTelemetry::RINIT::Loading .env file%A
 [%s] [DEBUG] [pid=%d] [ThreadId(%d)] otel::trace::tracer_provider: event src/trace/tracer_provider.rs:%d message=creating tracer provider for key (%d, "from-dotenv:service.namespace=my-dotenv-service,service.version=0.1.0")
 %A
 Spans
